Extract send bitcoins form to separate file
Extracts `SendBitcoinsToDepositAddressForm` to a separate file to make the
`MakeDeposit` file cleaner.
