+++ src/components/tBTC/index.ts
@@ -4,3 +4,4 @@ export * from "./Stats"
 export * from "./tBTCText"
 export * from "./BridgeActivity"
 export * from "./BridgeProcessIndicator"
+export * from "./SendBitcoinsToDepositAddressForm"
